This is an attempt to make the subtypes of `AbstractVariable` behave like `DiskArrays`. See JuliaGeo/CommonDataModel.jl#8.
I'm not perfectly sure of what I'm doing, but simple cases already work:

```julia
using NCDatasets
ds = NCDataset("tos_O1_2001-2002.nc")
var[:,:,1]
julia> var[:,:,1]
180×170 reshape(::Array{Union{Missing, Float32}, 3}, 180, 170) with eltype Union{Missing, Float32}:
 missing  missing  missing  missing  missing  missing  missing  missing  missing  …  271.419  271.42   271.422  271.426  271.43   271.437  271.445  271.459
 missing  missing  missing  missing  missing  missing  missing  missing  missing     271.419  271.42   271.422  271.426  271.431  271.438  271.445  271.459
 missing  missing  missing  missing  missing  missing  missing  missing  missing     271.418  271.42   271.422  271.425  271.431  271.438  271.445  271.459
 ⋮                                            ⋮                                   ⋱                               ⋮                                 
 missing  missing  missing  missing  missing  missing  missing  missing  missing     271.42   271.421  271.423  271.426  271.43   271.437  271.445  271.459
 missing  missing  missing  missing  missing  missing  missing  missing  missing     271.42   271.42   271.423  271.426  271.43   271.437  271.445  271.459
```

But of course it doesn't always work:
```julia
julia> var[2:4,1:3,1]
ERROR: BoundsError: attempt to access 3×3×1 Array{Union{Missing, Float32}, 3} at index [2:4, 1:3, 1:1]
Stacktrace:
  [1] throw_boundserror(A::Array{Union{Missing, Float32}, 3}, I::Tuple{UnitRange{Int64}, UnitRange{Int64}, UnitRange{Int64}})
    @ Base ./abstractarray.jl:691
  [2] checkbounds
    @ ./abstractarray.jl:656 [inlined]
  [3] view(::Array{Union{Missing, Float32}, 3}, ::UnitRange{Int64}, ::UnitRange{Int64}, ::UnitRange{Int64})
    @ Base ./subarray.jl:177
  [4] maybeview(::Array{Union{Missing, Float32}, 3}, ::UnitRange{Int64}, ::UnitRange{Int64}, ::Vararg{UnitRange{Int64}})
    @ Base ./views.jl:146
  [5] dotview(::Array{Union{Missing, Float32}, 3}, ::UnitRange{Int64}, ::UnitRange{Int64}, ::Vararg{UnitRange{Int64}})
    @ Base.Broadcast ./broadcast.jl:1200
  [6] readblock!(::NCDatasets.Variable{Float32, 3, NCDataset{Nothing}}, ::Array{Union{Missing, Float32}, 3}, ::UnitRange{Int64}, ::UnitRange{Int64}, ::Vararg{UnitRange{Int64}})
    @ NCDatasets ~/.julia/dev/NCDatasets/src/variable.jl:449
  [7] readblock!(::CommonDataModel.CFVariable{Union{Missing, Float32}, 3, NCDatasets.Variable{Float32, 3, NCDataset{Nothing}}, NCDatasets.Attributes{NCDataset{Nothing}}, NamedTuple{(:fillvalue, :missing_values, :scale_factor, :add_offset, :calendar, :time_origin, :time_factor), Tuple{Float32, Tuple{Float32}, Nothing, Nothing, Nothing, Nothing, Nothing}}}, ::Array{Union{Missing, Float32}, 3}, ::UnitRange{Int64}, ::UnitRange{Int64}, ::Vararg{UnitRange{Int64}})
    @ CommonDataModel ~/.julia/dev/CommonDataModel/src/cfvariable.jl:389
  [8] getindex_disk(::CommonDataModel.CFVariable{Union{Missing, Float32}, 3, NCDatasets.Variable{Float32, 3, NCDataset{Nothing}}, NCDatasets.Attributes{NCDataset{Nothing}}, NamedTuple{(:fillvalue, :missing_values, :scale_factor, :add_offset, :calendar, :time_origin, :time_factor), Tuple{Float32, Tuple{Float32}, Nothing, Nothing, Nothing, Nothing, Nothing}}}, ::UnitRange{Int64}, ::Vararg{Any})
    @ DiskArrays ~/.julia/packages/DiskArrays/zaCXL/src/diskarray.jl:31
  [9] getindex(::CommonDataModel.CFVariable{Union{Missing, Float32}, 3, NCDatasets.Variable{Float32, 3, NCDataset{Nothing}}, NCDatasets.Attributes{NCDataset{Nothing}}, NamedTuple{(:fillvalue, :missing_values, :scale_factor, :add_offset, :calendar, :time_origin, :time_factor), Tuple{Float32, Tuple{Float32}, Nothing, Nothing, Nothing, Nothing, Nothing}}}, ::UnitRange{Int64}, ::UnitRange{Int64}, ::Int64)
    @ DiskArrays ~/.julia/packages/DiskArrays/zaCXL/src/diskarray.jl:177
 [10] top-level scope
    @ REPL[17]:1
```

I will try to solve this in the coming days, any help is of course welcomed, and as I said above, I'm unsure about the approach I'm following :)

cc @rafaqz @Alexander-Barth