@rafaqz 2023-04-15
Hmm I dont get how your second example could ever work?