@Alexander-Barth 2023-04-15
Unfortunately, we still have an infinite recursion when loading a scalar:

https://github.com/Alexander-Barth/NCDatasets.jl/blob/pr-DiskArray/test/test_scalar.jl#L16