@rafaqz 2023-04-15
Indexing a muktidimensional array with a Colon like that is again breaking the base julia the array interface ;)