@Alexander-Barth 2023-04-15
the data in question is a zero-dimensional array. This does work in base julia:

```julia
zero_dim_array =  dropdims([1],dims=1)
zero_dim_array[:]
```

`[:]` just flattens the array:

```
julia> ones(2,2)[:]
4-element Vector{Float64}:
 1.0
 1.0
 1.0
 1.0
```

For some reasons my commits do not show in this PR, strange.