@rafaqz 2023-04-15
Yes it should work, returning a vector.

But your testing that the result is a scalar?