@rafaqz 2023-04-17
If you want to add diskarrays behavioir just to variables, you can do that using the macros provided in DiskArrays.jl, rather than suntyping AbstractDiskArray.

But Im not sure you would want any disk based array types to not have disk arrays behaviour, lazy chunked broadcasting is a pretty nice thing to have, instead if broadcasts just not working at all.