@Alexander-Barth 2023-04-17
I am wondering if (at least as a start) we just convert the base type NCDatasets.Variable to derive from DiskArray. 
In this case, the conversion from Array{Union{Missing,T},N} to Array{T,N} is already taken care of.

Maybe the macros that @rafaqz mentioned can then be used for wrapped types like CFVariable without requiring that they derive from DiskArray?

You guys have a better understand of DiskArrays works... Is there any documentation about the macros? 

> Strange indeed! Do you commit to my forked diskarrays branch? Maybe I have to setup something in my forked repo?

I used the commands from https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/reviewing-changes-in-pull-requests/checking-out-pull-requests-locally to test the PR:

```
git fetch origin pull/205/head:pr-DiskArray
git checkout pr-DiskArray 
```

and I assumed that any commits with show up here, but maybe that work only for PR that I initiate. In any case, if things become simplier for you, if you are a dev on this repo just let me know.

@rafaqz In general, yes some the test values should change. The issue of the recursion happened before but maybe it is related to this:

https://github.com/Alexander-Barth/NCDatasets.jl/pull/205/files#diff-fe41cd4463f1540b047828f14c99e312c078e817112d8f201a49eace1f5d58b9R495

I will already change the test in the master version.