@tcarion 2023-04-17
I tried to run the test and I spotted a few general issues:

- When scalar variables are implied
- When infinite dimensions are implied
- When writing to the `Variable` with a different eltype ([e.g. here](https://github.com/Alexander-Barth/NCDatasets.jl/blob/690ba2206f5cd7f5395f9f92ad3d071ead950350/test/test_scaling.jl#L33))
- When dealing with other subtypes of `CommonDataModel.AbstractVariable` (`SubVariable`, `MFVariable`). That would imply to change all the `getindex` / `setindex!` methods with their `readblock!`/`writeblock` equivalent. That's why it might be more simple to add diskarrays behaviour only to Variable as you suggested above, but I don't know how to do this.