@rafaqz 2023-04-17
See here for disk arrays macros. Above that application you can see all the sub macros. But you probably want the main one;

https://github.com/meggart/DiskArrays.jl/blob/main/src/DiskArrays.jl#L37



We could run your tests in CI so we can all see the failures and discuss them? I think thats the best way for me to contribute to the discussion.

@tcarion I will make you a JuliaGeo member so you dont need run permissions. You already have a package here anyway.