@Alexander-Barth 2023-04-18
Concerning the error type "2. When infinite dimensions are implied" there are some changes API, for example when previously we could do 

```
# ncvar is a NetCDF vector variable with an ulimited dimension
ncvar[:] = data
```
we now need to do:

```
# ncvar is a NetCDF vector variable with an ulimited dimension
ncvar[1:length(data)] = data
```

which mimics the behaviour of NetCDF.jl.

For the other cases, however, I don't think (or hope :-) ) that we need to break current API.

Another point is that `ncvar[:]` currently loads all the data and preserving the shape (without flatting the array as in julia). This behavior can go away too. I tried to remove tests that rely on this behaviour, but I might have forgotten some.

Thanks @rafaqz for the info!

> So @implement_diskarray Variable should do it.

To be certain that I understand: we can either derive Variable from AbstractDiskArrays or we can use the macro to inherit the behavior (chunking,...) . Is this correct? Or do we need to do both?

What do you think about the approach to make just NCDatasets.Variable as a DiskArrays and all types wrapping NCDatasets.Variable would use method forwarding as [explained here](https://discourse.julialang.org/t/method-forwarding-macro/23355/4)?


I changed the setting in github action; I hope it works now. (If not maybe @tcarion can make a trivial change the to README.md so that he is not considered as a first time contributor to NCDatasets for github.)