@Alexander-Barth 2023-08-04
Unfortunately, I see some slow down for the benchmark on my machine. Current master runs in 181.193 ms  while this PR 307.836 ms. Do you see this as well? I am using DiskArrays v0.3.14 (current version).

I tried this PR with another package of our package (DIVAnd) and upgrade was quite smooth.

For info: this commit  7edf68f7891fb28dba5a92acc6cbd702509f386e disable the opendap tests which causes current test failures (opendap server seems to be down).


Master
```
julia> bm
BenchmarkTools.Trial: 100 samples with 1 evaluation.
 Range (min … max):  162.789 ms … 231.338 ms  ┊ GC (min … max): 1.82% … 15.08%
 Time  (median):     181.193 ms               ┊ GC (median):    2.68%
 Time  (mean ± σ):   183.786 ms ±  16.540 ms  ┊ GC (mean ± σ):  3.09% ±  1.54%

  █ ▄▂▂            ▂▆  ▂ ▂ ▂                                     
  █▆███▆▆▁▄▆▆▆▆█▄▄███▄▄█▄█▁███▆▆▆▆▁▆▄▁▄▁▁▁▄▄▁▁▄▄▁▄▁▁▁▁▁▁▆▄▄▄▆▁▄ ▄
  163 ms           Histogram: frequency by time          224 ms <

 Memory estimate: 476.90 MiB, allocs estimate: 1610.

```

PR
```
julia> bm
BenchmarkTools.Trial: 100 samples with 1 evaluation.
 Range (min … max):  291.086 ms … 326.860 ms  ┊ GC (min … max): 1.15% … 1.50%
 Time  (median):     307.836 ms               ┊ GC (median):    1.74%
 Time  (mean ± σ):   307.185 ms ±   4.399 ms  ┊ GC (mean ± σ):  1.74% ± 0.22%

                            ▃   ▃ ▂     ▅▂▅ █▃                   
  ▄▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▄▁▁▄▁▅▄▅▄█▁█▇█▄█▅▄▄▇▇███▇████▅▄▇▄▄▁▁▁▁▁▁▁▁▁▄ ▄
  291 ms           Histogram: frequency by time          318 ms <

 Memory estimate: 527.15 MiB, allocs estimate: 8010.
```